+++ Demo.MultipleApis.Tests/Usings.cs
@@ -1,0 +1,5 @@
+global using rbkApiModules.Commons.Testing;
+global using rbkApiModules.Testing.Core;
+global using Shouldly;
+global using System.Net;
+global using System.Text.Json;