+++ Demo1.Tests/Demo1TestingServer.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace Demo1.Tests;
 
@@ -11,10 +9,6 @@ public class Demo1TestingServer : RbkTestingServer<Program>
     protected override Task InitializeApplicationAsync()
     {
         return Task.CompletedTask;
-    }
-
-    protected override void ConfigureAppConfiguration(WebHostBuilderContext context, IConfigurationBuilder config)
-    {
     }
 
     protected override void ConfigureTestServices(IServiceCollection services)