+++ Demo2.Tests/Demo2TestingServer.cs
@@ -1,5 +1,3 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Demo2.Tests;
@@ -11,10 +9,6 @@ public class Demo2TestingServer : RbkTestingServer<Demo2.Program>
     protected override Task InitializeApplicationAsync()
     {
         return Task.CompletedTask;
-    }
-
-    protected override void ConfigureAppConfiguration(WebHostBuilderContext context, IConfigurationBuilder config)
-    {
     }
 
     protected override void ConfigureTestServices(IServiceCollection services)