+++ Demo2/Clients/IDemo1ApiClient.cs
@@ -1,0 +1,6 @@
+namespace Demo2.Clients;
+
+public interface IDemo1ApiClient
+{
+    Task<string> GetAnonymousMessageAsync(CancellationToken cancellationToken = default);
+}